+++ HealthBars/Framework/ModConfig.cs
@@ -1,3 +1,4 @@
+using System.Runtime.Serialization;
 using Microsoft.Xna.Framework;
 using StardewModdingAPI;
 using StardewModdingAPI.Utilities;
@@ -7,6 +8,9 @@ namespace HealthBars.Framework;
 /// <summary>The mod configuration.</summary>
 internal class ModConfig
 {
+    /*********
+    ** Accessors
+    *********/
     /// <summary>Whether to show a health bar for monsters at full health.</summary>
     public bool DisplayHealthWhenNotDamaged { get; set; }
 
@@ -45,4 +49,16 @@ internal class ModConfig
 
     /// <summary>The keys which reload the mod config.</summary>
     public KeybindList ReloadKey { get; set; } = new(SButton.F5);
+
+
+    /*********
+    ** Private methods
+    *********/
+    /// <summary>The method called after the config file is deserialized.</summary>
+    /// <param name="context">The deserialization context.</param>
+    [OnDeserialized]
+    private void OnDeserializedMethod(StreamingContext context)
+    {
+        this.ReloadKey ??= new KeybindList(SButton.None);
+    }
 }