+++ MovementMod/Framework/ModConfig.cs
@@ -1,3 +1,4 @@
+using System.Runtime.Serialization;
 using StardewModdingAPI;
 using StardewModdingAPI.Utilities;
 
@@ -26,4 +27,17 @@ internal class ModConfig
 
     /// <summary>The stamina drain each second while sprinting.</summary>
     public float SprintingStaminaDrainPerSecond { get; set; } = 15;
+
+
+    /*********
+    ** Private methods
+    *********/
+    /// <summary>The method called after the config file is deserialized.</summary>
+    /// <param name="context">The deserialization context.</param>
+    [OnDeserialized]
+    private void OnDeserializedMethod(StreamingContext context)
+    {
+        this.SprintKey ??= new KeybindList(SButton.None);
+        this.ReloadKey ??= new KeybindList(SButton.None);
+    }
 }