+++ HealthBars/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 ## Release notes
+## Upcoming release
+* Fixed error if the `ReloadKey` config option is set to `null`.
+
 ## 2.8.5
 Released 04 November 2024 for SMAPI 4.1.2 or later. Updated by Pathoschild.